server/pyproject.toml
[project]
name = "pickfestar"
version = "0.1.0"
description = ""
authors = [
    {name = "jykim",email = "alcy1024@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "qdrant-client (>=1.14.3,<2.0.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "ollama (>=0.5.1,<0.6.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "langchain-ollama (>=0.3.3,<0.4.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "langchain-huggingface (>=0.3.0,<0.4.0)",
    "langchain-qdrant (>=0.2.0,<0.3.0)"
]

[tool.poetry]
packages = [{include = "src", from = "."}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
